created unique values contraints for username and role

+++ book_review/src/main/java/dev/abibou/bookreview/repository/UserRepository.java
@@ -1,13 +1,17 @@
 package dev.abibou.bookreview.repository;
 
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import dev.abibou.bookreview.entity.UserEntity;
 
 @Repository
-public interface UserRepository extends CrudRepository<UserEntity, Long> {
-	
+public interface UserRepository extends CrudRepository<UserEntity, Integer> {
+
 	UserEntity findByUsername(String username);
+
+//	@Query("delete from users where username = :username")
+//	void deleteByUsername(String username);
 
 }
